Refactor chatwidget input flow into modules
